+++ R/objects.R
@@ -286,7 +286,7 @@ gcs_get_object <- function(object_name,
 #' Use this to pass to uploads in \link{gcs_upload}
 #'
 #' @inheritParams Object
-#' @param object_name Name of the object. GCS uses this version if also set elsewhere.
+#' @param object_name Name of the object. GCS uses this version if also set elsewhere, or a \code{gs://} URL
 #'
 #' @return Object metadata for uploading of class \code{gar_Object}
 #' @family object functions
@@ -325,7 +325,7 @@ gcs_metadata_object <- function(object_name = NULL,
 #'
 #' Deletes an object from a bucket
 #'
-#' @param object_name Object to be deleted
+#' @param object_name Object to be deleted, or a \code{gs://} URL
 #' @param bucket Bucket to delete object from
 #' @param generation If present, deletes a specific version.
 #'
@@ -369,6 +369,68 @@ gcs_delete_object <- function(object_name,
   myMessage("Deleted '", object_name, "' from bucket '", bucket,"'")
   TRUE
 
+}
+
+#' Copy an object
+#' 
+#' Copies an object to a new destination
+#' 
+#' @param source_object The name of the object to copy, or a \code{gs://} URL
+#' @param destination_object The name of where to copy the object to, or a \code{gs://} URL
+#' @param source_bucket The bucket of the source object
+#' @param destination_bucket The bucket of the destination
+#' @param rewriteToken Include this field (from the previous rewrite response) on each rewrite request after the first one, until the rewrite response 'done' flag is true.
+#' @param destinationPredefinedAcl Apply a predefined set of access controls to the destination object. If not NULL must be one of the predefined access controls such as \code{"bucketOwnerFullControl"}
+#' 
+#' @return If successful, a rewrite object.
+#' @family object functions
+#' @import assertthat
+#' @importFrom googleAuthR gar_api_generator
+#' @export
+gcs_copy_object <- function(source_object,
+                            destination_object,
+                            source_bucket = gcs_get_global_bucket(),
+                            destination_bucket = gcs_get_global_bucket(),
+                            rewriteToken = NULL,
+                            destinationPredefinedAcl = NULL){
+  
+  assert_that(
+    is.string(source_object),
+    is.string(destination_object)
+  )
+  
+  source_object <- gcs_parse_gsurls(source_object)
+  if(!is.null(source_object)){
+    source_object <- source_object$obj
+    source_bucket <- source_object$bucket
+  }
+  destination_object <- gcs_parse_gsurls(destination_object)
+  if(!is.null(destination_object)){
+    destination_object <- destination_object$obj
+    destination_bucket <- destination_object$bucket
+  }
+  
+  source_bucket <- as.bucket_name(source_bucket)
+  destination_bucket <- as.bucket_name(destination_bucket)
+  
+  source_object <- URLencode(source_object, reserved = TRUE)
+  destination_object <- URLencode(destination_object, reserved = TRUE)
+  
+  the_url <- sprintf("https://www.googleapis.com/storage/v1/b/%s/o/%s/rewriteTo/b/%s/o/%s", 
+                     source_bucket, source_object, destination_bucket, destination_object)
+  pars <- NULL
+  if(!is.null(rewriteToken)){
+    pars <- list(rewriteToken = rewriteToken)
+  }
+  if(!is.null(destinationPredefinedAcl)){
+    assert_that(is.string(destinationPredefinedAcl))
+    pars <- c(pars, list(destinationPredefinedAcl = destinationPredefinedAcl))
+  }
+  ob <- gar_api_generator(the_url,
+                          "POST",
+                          pars_args = pars,
+                          data_parse_function = function(x) x)
+  ob()
 }
 
 #' Object Object