+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Make `gcs_auth()` more forgiving
 * Add a default google cloud project
+* Prevent error if bucket name is empty string (#90)
+* `gcs_upload()` will use file extension of `name` in its temporary file (#91)
 
 # googleCloudStorageR 0.4.0