+++ NEWS.md
@@ -3,6 +3,7 @@
 * Refactor authentication to favour json file and fix auth hangs (#87)
 * Prevent error if bucket name is empty string (#90)
 * `gcs_upload()` will use file extension of `name` in its temporary file (#91)
+* Add `gcs_copy_object()`
 
 # googleCloudStorageR 0.4.0