+++ NEWS.md
@@ -4,6 +4,8 @@
 * Prevent error if bucket name is empty string (#90)
 * `gcs_upload()` will use file extension of `name` in its temporary file (#91)
 * Add `gcs_copy_object()`
+* Add `gcs_compose_objects()`
+* Refactor `gcs_list_objects()` to googleAuthR > 0.7 `gar_api_page()`
 
 # googleCloudStorageR 0.4.0